\name{stratisamp-package}
\alias{stratisamp-package}
\alias{stratisamp}
\docType{package}
\title{
\packageTitle{stratisamp}
}
\description{
\packageDescription{stratisamp}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{stratisamp}
\packageIndices{stratisamp}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{stratisamp}

Maintainer: \packageMaintainer{stratisamp}
}
\references{
https://gist.githubusercontent.com/mrdwab/984691/raw/34689cfc35e636210eb16d3750be5b09e662582a/Stratified.R
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
}
\examples{
}
